pyproject.toml
[project]
name = "learning-managment-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"

requires-python = "^3.13"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "redis (>=5.2.1,<6.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "punq (>=0.7.0,<0.8.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "aiosmtplib (>=4.0.0,<5.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
